requirements.txt
google-api-python-client==2.187.0
Django>=5.2,<6.0
requests>=2.31,<3
imaplib2==3.6
PyDrive2==1.21.3
selenium>=4.15
undetected-chromedriver>=3.5
python-dotenv>=1.0
django-unfold==1.0.0